+++ test/stabilization_engine_band_rejections_test.dart
@@ -7,7 +7,7 @@ import 'package:ocr_stabilizer/src/band_fallback_config.dart';
 import 'package:ocr_stabilizer/src/default_tracked_block.dart';
 import 'package:ocr_stabilizer/src/stabilization_engine.dart';
 import 'package:ocr_stabilizer/src/stabilizer_config.dart';
-import 'package:ocr_stabilizer/src/tracked_block.dart';
+import 'package:ocr_stabilizer/src/observation.dart';
 import 'package:ocr_stabilizer/src/types/absolute_rect.dart';
 
 DefaultTrackedBlock<Object> _block(String text,
@@ -55,7 +55,7 @@ void main() {
             bandFallback: BandFallbackConfig(
               mode: BandFallbackMode.admit,
               candidateObservationFloor: 1,
-              spatialConfirm: (TrackedBlock a, TrackedBlock b) => false,
+              spatialConfirm: (Observation a, Observation b) => false,
             ),
           ),
         ),